Split resolver metadata queue tests
