+++ src/typechecker/tests/resolver_metadata.rs
@@ -4,5 +4,6 @@ mod formatting_and_behavior_refs;
 mod impl_and_method_helpers;
 mod metadata_requirements;
 mod metadata_restoration;
+mod queue_selection;
 mod type_and_key_helpers;
 mod validation_descriptors;
